encoder value is an integer instead of a float

+++ src/Drive/Drive.h
@@ -23,7 +23,7 @@ public:
 	float getForwardSpeed();
 	float getTurnSpeed();
 
-	float getCANTalonEncPos();
+	int getCANTalonEncPos();
 	float getCANTalonEncVel();
 
 private:
